requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
asgiref            # via django
attrs            # via pytest
certifi       # via requests
chardet            # via requests
django             # via -r requirements.in, djangorestframework
djangorestframework  # via -r requirements.in
idna         # via requests
importlib-metadata  # via pluggy, pytest
more-itertools     # via pytest
packaging          # via pytest
pluggy           # via pytest
py                # via pytest
pyparsing          # via packaging
pytest-django      # via -r requirements.in
pytest             # via pytest-django
pytz             # via django
requests        # via -r requirements.in
six              # via packaging
sqlparse           # via django
urllib3           # via requests
wcwidth           # via pytest
zipp               # via importlib-metadata
djangorestframework-simplejwt
django-filter